added non-interactive grocery list
