+++ src/components/savedRecipes/SavedRecipes.css
@@ -42,7 +42,7 @@
     display: block;
     margin-left: auto;
     margin-right: auto;
-    min-width: 75%;
+    min-width: 70%;
 
 }
